fix: address sage round-3 findings on #338
Two findings applied, one deferred:

1. **Architecture (important): runtime API depended on review module.**
   Pre-fix the exported `MyelinRuntime.jetstreamManager` type pointed at
   `bus/jetstream/provision.ts` — the review-specific module — making
   the generic bus runtime structurally depend on review provisioning
   concerns. Extract the narrow `ProvisionJsm` interface into a new
   neutral `bus/jetstream/types.ts`; both `provision.ts` and `runtime.ts`
   import the type from there. Provision module re-exports it so
   existing callers see no surface change.

2. **CodeQuality (suggestion): set-compare subjects in drift detection.**
   JetStream stream subjects are semantically a set — order on the wire
   is implementation-detail. Pre-fix `describeStreamDrift` used
   index-by-index comparison, which would false-warn on every boot for
   a live config that returned the same subjects in a different order.
   Switch to set-comparison; warning message also surfaces both sides
   as sorted `{}` notation to make the comparison visible.

   New unit test pins the set-equality invariant.

3. **Skipped: parallelize per-agent durable provisioning.** Same defer
   as round 2 — savings are 100-500ms on a typical 4-agent boot;
   sequential per-agent error correlation already lives in the loop
   shape; defer until an operator reports the latency.

Verified locally: bunx tsc --noEmit clean, full suite 3379 pass / 0
fail / 2 skip (was 3378 → +1 new test).
